+++ sdk/resourcemanager/healthdataaiservices/armhealthdataaiservices/fake/deidservices_server.go
@@ -39,10 +39,6 @@ type DeidServicesServer struct {
 	// NewListBySubscriptionPager is the fake for method DeidServicesClient.NewListBySubscriptionPager
 	// HTTP status codes to indicate success: http.StatusOK
 	NewListBySubscriptionPager func(options *armhealthdataaiservices.DeidServicesClientListBySubscriptionOptions) (resp azfake.PagerResponder[armhealthdataaiservices.DeidServicesClientListBySubscriptionResponse])
-
-	// BeginUpdate is the fake for method DeidServicesClient.BeginUpdate
-	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted
-	BeginUpdate func(ctx context.Context, resourceGroupName string, deidServiceName string, properties armhealthdataaiservices.DeidUpdate, options *armhealthdataaiservices.DeidServicesClientBeginUpdateOptions) (resp azfake.PollerResponder[armhealthdataaiservices.DeidServicesClientUpdateResponse], errResp azfake.ErrorResponder)
 }
 
 // NewDeidServicesServerTransport creates a new instance of DeidServicesServerTransport with the provided implementation.
@@ -55,7 +51,6 @@ func NewDeidServicesServerTransport(srv *DeidServicesServer) *DeidServicesServer
 		beginDelete:                 newTracker[azfake.PollerResponder[armhealthdataaiservices.DeidServicesClientDeleteResponse]](),
 		newListByResourceGroupPager: newTracker[azfake.PagerResponder[armhealthdataaiservices.DeidServicesClientListByResourceGroupResponse]](),
 		newListBySubscriptionPager:  newTracker[azfake.PagerResponder[armhealthdataaiservices.DeidServicesClientListBySubscriptionResponse]](),
-		beginUpdate:                 newTracker[azfake.PollerResponder[armhealthdataaiservices.DeidServicesClientUpdateResponse]](),
 	}
 }
 
@@ -67,7 +62,6 @@ type DeidServicesServerTransport struct {
 	beginDelete                 *tracker[azfake.PollerResponder[armhealthdataaiservices.DeidServicesClientDeleteResponse]]
 	newListByResourceGroupPager *tracker[azfake.PagerResponder[armhealthdataaiservices.DeidServicesClientListByResourceGroupResponse]]
 	newListBySubscriptionPager  *tracker[azfake.PagerResponder[armhealthdataaiservices.DeidServicesClientListBySubscriptionResponse]]
-	beginUpdate                 *tracker[azfake.PollerResponder[armhealthdataaiservices.DeidServicesClientUpdateResponse]]
 }
 
 // Do implements the policy.Transporter interface for DeidServicesServerTransport.
@@ -103,8 +97,6 @@ func (d *DeidServicesServerTransport) dispatchToMethodFake(req *http.Request, me
 				res.resp, res.err = d.dispatchNewListByResourceGroupPager(req)
 			case "DeidServicesClient.NewListBySubscriptionPager":
 				res.resp, res.err = d.dispatchNewListBySubscriptionPager(req)
-			case "DeidServicesClient.BeginUpdate":
-				res.resp, res.err = d.dispatchBeginUpdate(req)
 			default:
 				res.err = fmt.Errorf("unhandled API %s", method)
 			}
@@ -316,54 +308,6 @@ func (d *DeidServicesServerTransport) dispatchNewListBySubscriptionPager(req *ht
 	if !server.PagerResponderMore(newListBySubscriptionPager) {
 		d.newListBySubscriptionPager.remove(req)
 	}
-	return resp, nil
-}
-
-func (d *DeidServicesServerTransport) dispatchBeginUpdate(req *http.Request) (*http.Response, error) {
-	if d.srv.BeginUpdate == nil {
-		return nil, &nonRetriableError{errors.New("fake for method BeginUpdate not implemented")}
-	}
-	beginUpdate := d.beginUpdate.get(req)
-	if beginUpdate == nil {
-		const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceGroups/(?P<resourceGroupName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.HealthDataAIServices/deidServices/(?P<deidServiceName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
-		regex := regexp.MustCompile(regexStr)
-		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
-		if len(matches) < 4 {
-			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
-		}
-		body, err := server.UnmarshalRequestAsJSON[armhealthdataaiservices.DeidUpdate](req)
-		if err != nil {
-			return nil, err
-		}
-		resourceGroupNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceGroupName")])
-		if err != nil {
-			return nil, err
-		}
-		deidServiceNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("deidServiceName")])
-		if err != nil {
-			return nil, err
-		}
-		respr, errRespr := d.srv.BeginUpdate(req.Context(), resourceGroupNameParam, deidServiceNameParam, body, nil)
-		if respErr := server.GetError(errRespr, req); respErr != nil {
-			return nil, respErr
-		}
-		beginUpdate = &respr
-		d.beginUpdate.add(req, beginUpdate)
-	}
-
-	resp, err := server.PollerResponderNext(beginUpdate, req)
-	if err != nil {
-		return nil, err
-	}
-
-	if !contains([]int{http.StatusOK, http.StatusAccepted}, resp.StatusCode) {
-		d.beginUpdate.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusAccepted", resp.StatusCode)}
-	}
-	if !server.PollerResponderMore(beginUpdate) {
-		d.beginUpdate.remove(req)
-	}
-
 	return resp, nil
 }