+++ sdk/resourcemanager/healthdataaiservices/armhealthdataaiservices/responses.go
@@ -33,12 +33,6 @@ type DeidServicesClientListBySubscriptionResponse struct {
 	DeidServiceListResult
 }
 
-// DeidServicesClientUpdateResponse contains the response from method DeidServicesClient.BeginUpdate.
-type DeidServicesClientUpdateResponse struct {
-	// A HealthDataAIServicesProviderHub resource
-	DeidService
-}
-
 // OperationsClientListResponse contains the response from method OperationsClient.NewListPager.
 type OperationsClientListResponse struct {
 	// A list of REST API operations supported by an Azure Resource Provider. It contains an URL link to get the next set of results.